server/requirements.txt
fastapi
uvicorn[standard]
google-auth
google-api-python-client
pdfminer.six
pydantic
python-multipart
